// src/types/api.generated.ts
export interface components {
    schemas: {
// ...
        CreateWorkoutPlanDto: {
            /** @example Minha Ficha Hipertrofia ABCD */
            name: string;
            /** @example Ficha focada em braços e peitoral */
            description?: string;
            /** @enum {string} */
            goal?: "HYPERTROPHY" | "WEIGHT_LOSS" | "STRENGTH" | "CONDITIONING" | "HEALTH" | "OTHER";
            /** @enum {string} */
            difficulty?: "BEGINNER" | "INTERMEDIATE" | "ADVANCED";
            /** @example 4 */
            daysPerWeek: number;
            days: components["schemas"]["CreateWorkoutDayDto"][];
        };
// ...
    };
    responses: never;
    parameters: never;
    requestBodies: never;
    headers: never;
    pathItems: never;
}